Use system bus instead of session bus
Makes more sense in a service that should be available to whole system. But
also, NodeStateManager built with SimpleNodeStateMachine fails to start
when run as a service and outputs the following error message:

pconnection error: Cannot autolaunch D-Bus without X11 $DISPLAY

Since there is no session bus GLib tries to autolaunch one and it fails because
the DISPLAY environment variable is not set. This is per design, autolaunching a
session bus only makes sense if there is an X11 DISPLAY. See
0f2e4fd01c5cb07039aaf29f4f3219582c0f145c in GLib for more details:

https://gitlab.gnome.org/GNOME/glib/commit/0f2e4fd01c5cb07039aaf29f4f3219582c0f145c

Add a configuration file with same policy as for node-state-manager.

Signed-off-by: Martin Ejdestig <mejdestig@luxoft.com>